third_party/blink/web_tests/fast/events/pageshow-pagehide-on-back-uncached-expected.txt

ALERT: window.onload
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE ERROR: Blocked alert('window.onpagehide, target = [object HTMLDocument], persisted = false') during pagehide.
ALERT: window.onload
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = false
Test pageshow/pagehide event behavior when navigating back to an uncached page.
